code/tests/robot/requirements.txt
robotframework==7.1
robotframework-requests==0.9.7
robotframework-jsonlibrary==0.5
PyYAML==6.0.2
PyJWT==2.9.0
